Store current user in async storage and retrieve defaut garden
This commit makes it so that the current user is stored on sign in and
sign up. This makes it possible to retrieve the current user object in
other places. On the plant list, the current user is retrieved so the
user can be shown their default garden, rather than the previously
hard-coded garden.