Populate drawer with user's gardens
This commit updates the drawer menu so it is dynamically populated with
the users owned and shared gardens. Clicking on the name of a garden
pulls in that garden's plants from the API.

The menu was refactored a bit so it could wrap whichever view state is
being rendered, i.e., when there are plants or when there are no plants.