Add ImageWithIndicator
WHY

Images leave blank white spaces while loading.

WHAT

This commit adds a new component that adds a loading state to images so
that a user knows an image is loading.

When the image is initially loading, the ActivityIndicator is shown in
the center of where the image should appear. Once the image is finished
loading, the ActivityIndicator is hidden from view.

This looks like it works well with FastImage's `onLoadEnd` prop. Once
the image is cached, the ActivityIndicator doesn't seem to appear.