Add full size image view to check-in photos
This commit adds the ability for a user to press on of the images on a
check in and view the full sized image.

To get this working for Android, I had to introduce a top-level
`ImageViewProvider`. On Android, the `ImageView` would be stuck within
the bounds of it's containing element (which would have been the
`CheckInListItem`), so I placed the new provider at the top of the
hierarchy so it can take up the full screen.