Delete plants and refresh on return to plant list
This commit adds a new action button to the plant details screen that
allows a user to delete the plant.

I noticed while working on this that the plant list doesn't
automatically refresh after a plant is deleted (or added, edited, etc.),
so I added a new navigation focus listener. I wrapped this in the
`useDidFocus` hook so it's reusable on any screen where you want stuff
to happen when navigating back to a screen.

This commit also cleans up the code on the plant list screen. The logic
for re-fetching the current user and garden is a lot more
straightforward and more neatly encapsulated in `useEffect` hooks.

Finally I added a new jsonOrNull function that makes it easy to handle
an OK response that doesn't have a body.